refactor: split four large monolith files into sub-packages
Break up four files exceeding 1400 lines each so future agent
searches can read just the relevant sub-file instead of the
whole monolith. Public APIs are preserved verbatim via
re-export shims; no behavior changes.

- src/running_process/pty.py (2671 lines) -> pty/ package (12 files);
  largest is _pseudo_terminal.py (1647 lines, single class)
- src/running_process/running_process.py (1413 lines)
  -> running_process/ subpackage (6 files); largest is _core.py
  (1155 lines, single class)
- tests/test_pty_support.py (2540 lines) -> tests/pty/
  (10 feature-themed test files + _pty_helpers.py + conftest.py)
- crates/running-process-daemon/src/handlers.rs (1591 lines)
  -> handlers/ module dir (11 files), grouped by handler category;
  pub use re-exports in mod.rs keep crate::handlers::handle_* paths
  unchanged

Verified:
- uv run ruff check src tests ci -> clean
- soldr cargo check --workspace --all-targets -> clean (only
  pre-existing warnings in tree_kill_test)
- pytest --collect-only via the project's CLI wrapper -> 429
  tests collected with no errors
