+++ src/running_process/pty/_pseudo_terminal.py
@@ -1,34 +1,25 @@
 from __future__ import annotations
 
-import inspect
 import os
-import re
-import shlex
 import signal
 import sys
 import threading
 import time
-import warnings
 import weakref
 from collections.abc import Callable, Mapping
 from contextlib import suppress
-from dataclasses import dataclass, field, replace
-from enum import Enum
+from dataclasses import replace
 from io import TextIOBase
 from pathlib import Path
-from typing import Any, Literal
+from typing import Any
 
 from running_process._native import (
     NativeIdleDetector,
     NativeProcess,
     NativeProcessMetrics,
     NativePtyBuffer,
-    NativeSignalBool,
     NativeTerminalInput,
-    native_apply_process_nice,
 )
-from running_process.command_render import list2cmdline as render_command_list
-from running_process.compat import CREATE_NEW_PROCESS_GROUP
 from running_process.console_encoding import detect_console_encoding, sanitize_for_encoding
 from running_process.exit_status import ExitStatus, ProcessAbnormalExit, classify_exit_status
 from running_process.expect import (
@@ -40,509 +31,59 @@ from running_process.expect import (
     search_expect_pattern,
 )
 from running_process.priority import CpuPriority, normalize_nice
+from running_process.pty._command import _normalize_command, _pty_command, interactive_launch_spec
+from running_process.pty._console_io import _safe_console_write_chunk
+from running_process.pty._errors import PtyNotAvailableError, SignalBool
+from running_process.pty._idle_helpers import (
+    _build_default_idle_reset,
+    _compile_idle_detector,
+    _flush_wait_input,
+    _input_contains_newline,
+    _invoke_condition_callback,
+    _invoke_wait_callback,
+    _normalize_wait_conditions,
+    _resolve_expect_offset,
+    _start_event_count,
+)
+from running_process.pty._idle_state import (
+    _ExpectRuntimeState,
+    _IdleRuntimeState,
+    _IdleSample,
+    _WaitCallbackState,
+)
+from running_process.pty._process_helpers import (
+    _close_native_pty_process,
+    _warn_pty_text_mode_ignored,
+)
+from running_process.pty._types import (
+    Callback,
+    Expect,
+    Idle,
+    IdleContext,
+    IdleDecision,
+    IdleDetection,
+    IdleDetector,
+    IdleInfoDiff,
+    IdleReachedCallback,
+    IdleResetPredicate,
+    IdleStartTrigger,
+    IdleTiming,
+    IdleWaitResult,
+    InteractiveMode,
+    InterruptResult,
+    ProcessIdleDetection,
+    PtyIdleDetection,
+    WaitCallbackResult,
+    WaitCheckpoint,
+    WaitCondition,
+    WaitForResult,
+)
 
 _SUPPORTED_PTY_PLATFORMS = {"win32", "linux", "darwin"}
 _PTY_READ_CHUNK_TIMEOUT_SECONDS = 0.01
 _PTY_POLL_INTERVAL_SECONDS = 0.001
 _PTY_READER_NATIVE_CLOSE_WAIT_SECONDS = 2.0
 _PTY_CLEANUP_ERRORS = (OSError, RuntimeError, TimeoutError, ValueError, AttributeError)
-_NO_PTY_TEXT_WARNING_ENV = "RUNNING_PROCESS_NO_PTY_TEXT_WARNING"
-_WINDOWS_VT_OUTPUT_HANDLES: set[int] = set()
-_WINDOWS_VT_OUTPUT_LOCK = threading.Lock()
-
-
-class PtyNotAvailableError(RuntimeError):
-    pass
-
-
-class SignalBool:
-    def __init__(self, value: bool = False) -> None:
-        self._value = bool(value)
-        self._native = NativeSignalBool(self._value)
-
-    @property
-    def value(self) -> bool:
-        return self._value
-
-    @value.setter
-    def value(self, value: bool) -> None:
-        self._value = bool(value)
-        self._native.value = self._value
-
-    def load(self) -> bool:
-        return self._native.load_nolock()
-
-    def store(self, value: bool) -> None:
-        self.value = value
-
-    def compare_and_swap(self, current: bool, new: bool) -> bool:
-        swapped = self._native.compare_and_swap_locked(bool(current), bool(new))
-        if swapped:
-            self._value = bool(new)
-        else:
-            self._value = self._native.load_nolock()
-        return swapped
-
-    def __bool__(self) -> bool:
-        return self._value
-
-
-def _safe_console_write(stream: TextIOBase, line: str | bytes) -> None:
-    text = line.decode("utf-8", errors="replace") if isinstance(line, bytes) else line
-    _ensure_windows_vt_output(stream)
-    try:
-        stream.write(text)
-        stream.write("\n")
-    except UnicodeEncodeError:
-        encoding = stream.encoding or "utf-8"
-        rendered = text.encode(encoding, errors="replace")
-        if hasattr(stream, "buffer"):
-            stream.buffer.write(rendered + b"\n")
-        else:
-            stream.write(rendered.decode(encoding, errors="replace"))
-            stream.write("\n")
-    stream.flush()
-
-
-def _windows_console_output_handle(stream: TextIOBase) -> int | None:
-    if sys.platform != "win32":
-        return None
-    try:
-        fileno = stream.fileno()
-    except (AttributeError, OSError, ValueError):
-        return None
-    try:
-        import msvcrt
-    except ImportError:
-        return None
-    try:
-        return int(msvcrt.get_osfhandle(fileno))
-    except OSError:
-        return None
-
-
-def _enable_windows_vt_output_handle(handle: int) -> bool:
-    if sys.platform != "win32":
-        return False
-    try:
-        import ctypes
-    except ImportError:
-        return False
-
-    kernel32 = ctypes.windll.kernel32
-    mode = ctypes.c_uint32()
-    console_handle = ctypes.c_void_p(handle)
-    if kernel32.GetConsoleMode(console_handle, ctypes.byref(mode)) == 0:
-        return False
-
-    enable_processed_output = 0x0001
-    enable_virtual_terminal_processing = 0x0004
-    updated_mode = (
-        mode.value | enable_processed_output | enable_virtual_terminal_processing
-    )
-    if updated_mode == mode.value:
-        return True
-    return kernel32.SetConsoleMode(console_handle, updated_mode) != 0
-
-
-def _ensure_windows_vt_output(stream: TextIOBase) -> None:
-    handle = _windows_console_output_handle(stream)
-    if handle is None:
-        return
-    with _WINDOWS_VT_OUTPUT_LOCK:
-        if handle in _WINDOWS_VT_OUTPUT_HANDLES:
-            return
-        if _enable_windows_vt_output_handle(handle):
-            _WINDOWS_VT_OUTPUT_HANDLES.add(handle)
-
-
-def _safe_console_write_chunk(
-    stream: TextIOBase,
-    chunk: bytes,
-    *,
-    encoding: str,
-    errors: str,
-) -> None:
-    if not chunk:
-        return
-    _ensure_windows_vt_output(stream)
-    if hasattr(stream, "buffer"):
-        try:
-            stream.buffer.write(chunk)
-            stream.flush()
-            return
-        except UnicodeEncodeError:
-            pass
-    text = chunk.decode(encoding, errors)
-    try:
-        stream.write(text)
-    except UnicodeEncodeError:
-        fallback_encoding = stream.encoding or encoding or "utf-8"
-        rendered = text.encode(fallback_encoding, errors="replace")
-        if hasattr(stream, "buffer"):
-            stream.buffer.write(rendered)
-        else:
-            stream.write(rendered.decode(fallback_encoding, errors="replace"))
-    stream.flush()
-
-
-class InteractiveMode(str, Enum):
-    PSEUDO_TERMINAL = "pseudo_terminal"
-    CONSOLE_SHARED = "console_shared"
-    CONSOLE_ISOLATED = "console_isolated"
-
-
-@dataclass(frozen=True)
-class InteractiveLaunchSpec:
-    mode: InteractiveMode
-    uses_pty: bool
-    ctrl_c_owner: str
-    creationflags: int | None
-    restore_terminal: bool
-
-
-@dataclass(frozen=True)
-class InterruptResult:
-    exit_reason: str
-    interrupt_count: int
-    returncode: int | None
-
-
-class IdleDecision(str, Enum):
-    DEFAULT = "default"
-    ACTIVE = "active"
-    BEGIN_IDLE = "begin_idle"
-    IS_IDLE = "is_idle"
-
-
-class IdleStartTrigger(str, Enum):
-    IMMEDIATE = "immediate"
-    INPUT_NEWLINE = "input_newline"
-    INPUT_SUBMIT = "input_submit"
-
-
-@dataclass(slots=True)
-class IdleTiming:
-    timeout_seconds: float = 10.0
-    stability_window_seconds: float = 0.75
-    sample_interval_seconds: float = 0.25
-
-
-@dataclass(slots=True)
-class PtyIdleDetection:
-    reset_on_input: bool = True
-    reset_on_output: bool = True
-    count_control_churn_as_output: bool = True
-    start_trigger: IdleStartTrigger = IdleStartTrigger.IMMEDIATE
-
-
-@dataclass(slots=True)
-class ProcessIdleDetection:
-    cpu_percent_before_reset: float = 2.0
-    max_disk_io_bytes_before_reset: int = 4096
-    max_network_bytes_before_reset: int = 4096
-
-
-@dataclass(slots=True)
-class IdleInfoDiff:
-    delta_seconds: float
-    process_alive: bool
-    pty_input_bytes: int = 0
-    pty_output_bytes: int = 0
-    pty_control_churn_bytes: int = 0
-    cpu_percent: float = 0.0
-    disk_io_bytes: int = 0
-    network_io_bytes: int = 0
-
-    @property
-    def interval_seconds(self) -> float:
-        return self.delta_seconds
-
-
-@dataclass(slots=True)
-class IdleContext:
-    idle_for_seconds: float
-    stable_for_seconds: float
-    sample_count: int
-
-
-IdleDiff = IdleInfoDiff
-IdleReachedCallback = Callable[[IdleInfoDiff], IdleDecision]
-IdleResetPredicate = Callable[[IdleInfoDiff, IdleContext], bool]
-
-
-@dataclass(slots=True)
-class IdleDetection:
-    timing: IdleTiming = field(default_factory=IdleTiming)
-    pty: PtyIdleDetection | None = field(default_factory=PtyIdleDetection)
-    process: ProcessIdleDetection | None = None
-    idle_reached: IdleReachedCallback | None = None
-    predicate: IdleResetPredicate | None = None
-
-
-IdleDetector = IdleDetection | IdleResetPredicate | None
-
-
-@dataclass(frozen=True, slots=True)
-class IdleWaitResult:
-    returncode: int | None
-    idle_detected: bool
-    exit_reason: Literal["process_exit", "idle_timeout", "timeout", "interrupt"]
-    idle_for_seconds: float = 0.0
-
-    @property
-    def reason(self) -> str:
-        return self.exit_reason
-
-    @property
-    def idle_for(self) -> float:
-        return self.idle_for_seconds
-
-
-@dataclass(frozen=True, slots=True)
-class Idle:
-    detector: IdleDetector = field(default_factory=IdleDetection)
-    on_callback: Callable[..., object] | None = None
-
-
-class WaitCallbackResult(str, Enum):
-    EXIT = "exit"
-    CONTINUE = "continue"
-    CONTINUE_AND_DISARM = "continue_and_disarm"
-
-
-@dataclass(frozen=True, slots=True)
-class Expect:
-    pattern: ExpectPattern
-    action: ExpectAction = None
-    NOT: ExpectPattern | None = None
-    after: WaitCheckpoint | Literal["start", "now"] = "start"
-    on_callback: Callable[..., object] | None = None
-
-
-@dataclass(frozen=True, slots=True)
-class Callback:
-    callback: Callable[..., object]
-    poll_interval_seconds: float = 0.05
-
-
-WaitCondition = Idle | Expect | Callback
-
-
-@dataclass(frozen=True, slots=True)
-class WaitForResult:
-    returncode: int | None
-    matched: bool
-    exit_reason: Literal["condition_met", "process_exit", "timeout", "interrupt"]
-    condition: WaitCondition | None = None
-    expect_match: ExpectMatch | None = None
-    idle_result: IdleWaitResult | None = None
-    callback_result: object | None = None
-
-    @property
-    def reason(self) -> str:
-        return self.exit_reason
-
-
-@dataclass(frozen=True, slots=True)
-class WaitCheckpoint:
-    offset: int
-
-
-@dataclass(frozen=True, slots=True)
-class _BufferedInput:
-    data: str | bytes
-    submit: bool = False
-
-
-class WaitInputBuffer:
-    def __init__(self) -> None:
-        self._items: list[str | bytes | _BufferedInput] = []
-
-    def write(self, data: str | bytes) -> None:
-        self._items.append(data)
-
-    def submit(self, data: str | bytes) -> None:
-        self._items.append(_BufferedInput(data=data, submit=True))
-
-    def drain(self) -> list[str | bytes | _BufferedInput]:
-        items = list(self._items)
-        self._items.clear()
-        return items
-
-    def __bool__(self) -> bool:
-        return bool(self._items)
-
-
-@dataclass(slots=True)
-class _IdleRuntimeState:
-    last_reset_at: float
-    stable_since: float | None
-    sample_count: int = 0
-
-
-@dataclass(slots=True)
-class _IdleSample:
-    sampled_at: float
-    process_alive: bool
-    pty_input_bytes: int
-    pty_output_bytes: int
-    pty_control_churn_bytes: int
-    cpu_percent: float
-    disk_io_bytes: int
-    network_io_bytes: int
-    returncode: int | None
-
-
-@dataclass(slots=True)
-class _TerminalControlStripper:
-    mode: str = "capture"
-    _pending: bytearray = field(default_factory=bytearray)
-    _string_terminator: bytes | None = None
-
-    def strip(self, chunk: bytes) -> bytes:
-        if not chunk and not self._pending:
-            return b""
-        data = bytes(self._pending) + bytes(chunk)
-        self._pending.clear()
-        output = bytearray()
-        index = 0
-
-        while index < len(data):
-            if self._string_terminator is not None:
-                terminator = self._string_terminator
-                terminator_index = data.find(terminator, index)
-                if terminator_index == -1:
-                    self._pending.extend(data[index:])
-                    break
-                index = terminator_index + len(terminator)
-                self._string_terminator = None
-                continue
-
-            byte = data[index]
-            if byte == 0x1B:
-                if index + 1 >= len(data):
-                    self._pending.append(byte)
-                    break
-                marker = data[index + 1]
-                if marker == ord("["):
-                    end = _find_csi_end(data, index + 2)
-                    if end is None:
-                        self._pending.extend(data[index:])
-                        break
-                    normalized = _normalize_csi_sequence(
-                        data[index : end + 1],
-                        mode=self.mode,
-                    )
-                    if normalized:
-                        output.extend(normalized)
-                    index = end + 1
-                    continue
-                if marker == ord("]"):
-                    self._string_terminator = b"\x07"
-                    index += 2
-                    continue
-                if marker in {ord("P"), ord("X"), ord("^"), ord("_")}:
-                    self._string_terminator = b"\x1b\\"
-                    index += 2
-                    continue
-                index += 2
-                continue
-            if byte in {0x08, 0x7F}:
-                index += 1
-                continue
-            output.append(byte)
-            index += 1
-
-        return bytes(output)
-
-
-def _find_csi_end(data: bytes, start: int) -> int | None:
-    for index in range(start, len(data)):
-        current = data[index]
-        if 0x40 <= current <= 0x7E:
-            return index
-    return None
-
-
-_ECHO_DISPLAY_CSI_FINALS = {
-    b"A",
-    b"B",
-    b"C",
-    b"D",
-    b"E",
-    b"F",
-    b"G",
-    b"H",
-    b"J",
-    b"K",
-    b"S",
-    b"T",
-    b"f",
-    b"m",
-    b"s",
-    b"u",
-}
-
-
-def _normalize_csi_sequence(chunk: bytes, *, mode: str) -> bytes:
-    if len(chunk) < 3 or not chunk.startswith(b"\x1b["):
-        return b""
-    final = chunk[-1:]
-    params = chunk[2:-1]
-    if mode == "echo":
-        if final in _ECHO_DISPLAY_CSI_FINALS:
-            return b"\x1b[0m" if final == b"m" and params == b"" else chunk
-        if final in {b"h", b"l"}:
-            return chunk
-        return b""
-    if final == b"H" and params in {b"", b"1;1"}:
-        return b"\r"
-    if final == b"G" and params in {b"", b"1"}:
-        return b"\r"
-    return b""
-
-
-_TERMINAL_FRAGMENT_RE = re.compile(rb"(?:\d+;){2,}\d+_")
-
-
-def _strip_terminal_fragments(chunk: bytes) -> bytes:
-    if not chunk:
-        return chunk
-    return _TERMINAL_FRAGMENT_RE.sub(b"", chunk)
-
-
-def _collapse_duplicate_carriage_returns(chunk: bytes) -> bytes:
-    if not chunk:
-        return chunk
-    while b"\r\r\n" in chunk:
-        chunk = chunk.replace(b"\r\r\n", b"\r\n")
-    return chunk
-
-
-@dataclass(slots=True)
-class _IdleCallbackThreadState:
-    pending_diff: IdleInfoDiff | None = None
-    inflight: bool = False
-    latest_decision: IdleDecision | None = None
-    error: BaseException | None = None
-    closed: bool = False
-
-
-@dataclass(slots=True)
-class _WaitCallbackState:
-    ready: SignalBool = field(default_factory=SignalBool)
-    result: object | None = None
-    error: BaseException | None = None
-    pending_writes: list[str | bytes] = field(default_factory=list)
-    lock: threading.Lock = field(default_factory=threading.Lock)
-
-
-@dataclass(slots=True)
-class _ExpectRuntimeState:
-    search_offset: int = 0
-    armed: bool = True
-
 
 KEYBOARD_INTERRUPT_EXIT_CODES: set[int] = {
     -2,             # Unix: killed by SIGINT (negative signal number)
@@ -550,290 +91,6 @@ KEYBOARD_INTERRUPT_EXIT_CODES: set[int] = {
     -1073741510,    # Windows: STATUS_CONTROL_C_EXIT (signed)
     3221225786,     # Windows: STATUS_CONTROL_C_EXIT (unsigned)
 }
-
-
-def _compile_idle_detector(
-    idle_detector: IdleDetector,
-) -> tuple[IdleTiming | None, IdleReachedCallback | None, IdleResetPredicate | None]:
-    if idle_detector is None:
-        return None, None, None
-    if isinstance(idle_detector, IdleDetection):
-        if idle_detector.idle_reached is not None and idle_detector.predicate is not None:
-            raise ValueError("idle_reached and predicate are mutually exclusive")
-        if idle_detector.idle_reached is not None:
-            return idle_detector.timing, idle_detector.idle_reached, None
-        predicate = idle_detector.predicate or _build_default_idle_reset(idle_detector)
-        return idle_detector.timing, None, predicate
-    if callable(idle_detector):
-        callback_arity = _callable_arity(idle_detector)
-        if callback_arity == 1:
-            return IdleTiming(), idle_detector, None
-        if callback_arity == 2:
-            return IdleTiming(), None, idle_detector
-        raise TypeError("idle_detector callable must accept 1 or 2 positional arguments")
-    raise TypeError(
-        "idle_detector must be None, an IdleDetection instance, or a callable callback"
-    )
-
-
-def _callable_arity(callback: Callable[..., Any]) -> int:
-    signature = inspect.signature(callback)
-    required_positional = [
-        parameter
-        for parameter in signature.parameters.values()
-        if parameter.kind in (
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-        )
-        and parameter.default is inspect.Parameter.empty
-    ]
-    has_varargs = any(
-        parameter.kind is inspect.Parameter.VAR_POSITIONAL
-        for parameter in signature.parameters.values()
-    )
-    if has_varargs:
-        if len(required_positional) <= 1:
-            return 1
-        if len(required_positional) == 2:
-            return 2
-    if len(required_positional) in {1, 2}:
-        return len(required_positional)
-    raise TypeError("idle_detector callable must accept 1 or 2 positional arguments")
-
-
-def _wait_callback_arity(callback: Callable[..., object]) -> int:
-    signature = inspect.signature(callback)
-    required_positional = [
-        parameter
-        for parameter in signature.parameters.values()
-        if parameter.kind in (
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-        )
-        and parameter.default is inspect.Parameter.empty
-    ]
-    has_varargs = any(
-        parameter.kind is inspect.Parameter.VAR_POSITIONAL
-        for parameter in signature.parameters.values()
-    )
-    if has_varargs and len(required_positional) <= 2:
-        return len(required_positional)
-    if len(required_positional) in {0, 1, 2}:
-        return len(required_positional)
-    raise TypeError("wait callback must accept 0, 1, or 2 positional arguments")
-
-
-def _invoke_wait_callback(
-    callback: Callable[..., object], process: PseudoTerminalProcess
-) -> tuple[object, list[str | bytes]]:
-    arity = _wait_callback_arity(callback)
-    input_buffer = WaitInputBuffer()
-    if arity == 0:
-        result = callback()
-    elif arity == 1:
-        result = callback(input_buffer)
-    else:
-        result = callback(input_buffer, process)
-    return result, input_buffer.drain()
-
-
-def _condition_callback_arity(callback: Callable[..., object]) -> int:
-    signature = inspect.signature(callback)
-    required_positional = [
-        parameter
-        for parameter in signature.parameters.values()
-        if parameter.kind in (
-            inspect.Parameter.POSITIONAL_ONLY,
-            inspect.Parameter.POSITIONAL_OR_KEYWORD,
-        )
-        and parameter.default is inspect.Parameter.empty
-    ]
-    has_varargs = any(
-        parameter.kind is inspect.Parameter.VAR_POSITIONAL
-        for parameter in signature.parameters.values()
-    )
-    if has_varargs and len(required_positional) <= 3:
-        return len(required_positional)
-    if len(required_positional) in {0, 1, 2, 3}:
-        return len(required_positional)
-    raise TypeError("condition on_callback must accept 0, 1, 2, or 3 positional arguments")
-
-
-def _invoke_condition_callback(
-    callback: Callable[..., object],
-    payload: object,
-    process: PseudoTerminalProcess,
-) -> tuple[WaitCallbackResult, list[str | bytes]]:
-    arity = _condition_callback_arity(callback)
-    input_buffer = WaitInputBuffer()
-    if arity == 0:
-        result = callback()
-    elif arity == 1:
-        result = callback(payload)
-    elif arity == 2:
-        result = callback(payload, input_buffer)
-    else:
-        result = callback(payload, input_buffer, process)
-    if not isinstance(result, WaitCallbackResult):
-        raise TypeError("condition on_callback must return a WaitCallbackResult")
-    return result, input_buffer.drain()
-
-
-def _normalize_wait_conditions(
-    *conditions: (
-        WaitCondition
-        | Callable[..., object]
-        | list[WaitCondition | Callable[..., object]]
-        | tuple[WaitCondition | Callable[..., object], ...]
-    ),
-) -> list[WaitCondition]:
-    normalized: list[WaitCondition] = []
-    for condition in conditions:
-        if isinstance(condition, Idle | Expect | Callback):
-            normalized.append(condition)
-            continue
-        if callable(condition):
-            normalized.append(Callback(condition))
-            continue
-        if isinstance(condition, list | tuple):
-            for nested in condition:
-                if isinstance(nested, Idle | Expect | Callback):
-                    normalized.append(nested)
-                    continue
-                if callable(nested):
-                    normalized.append(Callback(nested))
-                    continue
-                raise TypeError("wait_for conditions must be Idle, Expect, Callback, or a callable")
-            continue
-        raise TypeError("wait_for conditions must be Idle, Expect, Callback, or a callable")
-    return normalized
-
-
-def _flush_wait_input(
-    process: PseudoTerminalProcess, items: list[str | bytes | _BufferedInput]
-) -> None:
-    for item in items:
-        if isinstance(item, _BufferedInput):
-            process.write(item.data, submit=item.submit)
-            continue
-        process.write(item)
-
-
-def _resolve_expect_offset(
-    condition: Expect, process: PseudoTerminalProcess
-) -> int:
-    if condition.after == "start":
-        return 0
-    if condition.after == "now":
-        return len(ensure_text(process.output, process.encoding, process.errors))
-    return max(0, condition.after.offset)
-
-
-def _build_default_idle_reset(cfg: IdleDetection) -> IdleResetPredicate:
-    return lambda diff, ctx: _default_idle_reset(diff, ctx, cfg)
-
-
-def _input_contains_newline(data: bytes) -> bool:
-    return b"\r" in data or b"\n" in data
-
-
-def _start_event_count(
-    process: PseudoTerminalProcess, start_trigger: IdleStartTrigger
-) -> int:
-    process._sync_native_input_metrics()
-    if start_trigger is IdleStartTrigger.INPUT_NEWLINE:
-        return process._pty_newline_events_total
-    if start_trigger is IdleStartTrigger.INPUT_SUBMIT:
-        return process._pty_submit_events_total
-    return 1
-
-
-def _default_idle_reset(diff: IdleDiff, _ctx: IdleContext, cfg: IdleDetection) -> bool:
-    pty_cfg = cfg.pty
-    if pty_cfg is not None:
-        if pty_cfg.reset_on_input and diff.pty_input_bytes > 0:
-            return True
-        output_bytes = diff.pty_output_bytes
-        if pty_cfg.count_control_churn_as_output:
-            output_bytes += diff.pty_control_churn_bytes
-        if pty_cfg.reset_on_output and output_bytes > 0:
-            return True
-
-    process_cfg = cfg.process
-    if process_cfg is not None:
-        if diff.cpu_percent > process_cfg.cpu_percent_before_reset:
-            return True
-        if diff.disk_io_bytes > process_cfg.max_disk_io_bytes_before_reset:
-            return True
-        if diff.network_io_bytes > process_cfg.max_network_bytes_before_reset:
-            return True
-
-    return False
-
-
-def _merge_idle_diff(base: IdleInfoDiff, update: IdleInfoDiff) -> IdleInfoDiff:
-    total_delta = base.delta_seconds + update.delta_seconds
-    weighted_cpu = 0.0
-    if total_delta > 0:
-        weighted_cpu = (
-            (base.cpu_percent * base.delta_seconds) + (update.cpu_percent * update.delta_seconds)
-        ) / total_delta
-    return IdleInfoDiff(
-        delta_seconds=total_delta,
-        process_alive=update.process_alive,
-        pty_input_bytes=base.pty_input_bytes + update.pty_input_bytes,
-        pty_output_bytes=base.pty_output_bytes + update.pty_output_bytes,
-        pty_control_churn_bytes=base.pty_control_churn_bytes + update.pty_control_churn_bytes,
-        cpu_percent=weighted_cpu,
-        disk_io_bytes=base.disk_io_bytes + update.disk_io_bytes,
-        network_io_bytes=base.network_io_bytes + update.network_io_bytes,
-    )
-
-
-def _control_churn_bytes(chunk: bytes) -> int:
-    total = 0
-    index = 0
-    while index < len(chunk):
-        byte = chunk[index]
-        if byte == 0x1B:
-            start = index
-            index += 1
-            if index < len(chunk) and chunk[index] == ord("["):
-                index += 1
-                while index < len(chunk):
-                    current = chunk[index]
-                    index += 1
-                    if 0x40 <= current <= 0x7E:
-                        break
-            total += index - start
-            continue
-        if byte in {0x08, 0x0D, 0x7F}:
-            total += 1
-        index += 1
-    return total
-
-
-def _close_native_pty_process(proc: NativeProcess | None) -> None:
-    if proc is None:
-        return
-    # Finalizers must not block indefinitely while the interpreter is collecting.
-    # Use best-effort non-blocking termination instead of `close()`.
-    with suppress(*_PTY_CLEANUP_ERRORS):
-        proc.kill()
-    with suppress(*_PTY_CLEANUP_ERRORS):
-        proc.terminate()
-
-
-def _warn_pty_text_mode_ignored(env: Mapping[str, str] | None) -> None:
-    effective_env = env if env is not None else os.environ
-    if effective_env.get(_NO_PTY_TEXT_WARNING_ENV):
-        return
-    warnings.warn(
-        "PTY mode ignores text/universal_newlines and always uses raw bytes; "
-        f"set {_NO_PTY_TEXT_WARNING_ENV}=1 to suppress this warning.",
-        RuntimeWarning,
-        stacklevel=3,
-    )
 
 
 class Pty:
@@ -2388,284 +1645,3 @@ class PseudoTerminalProcess:
             else:
                 wait_timeout = 0.05
             self._pump_native_output(timeout=wait_timeout, consume_all=True)
-
-
-class InteractiveProcess:
-    def __init__(
-        self,
-        command: str | list[str],
-        *,
-        mode: InteractiveMode | str = InteractiveMode.CONSOLE_SHARED,
-        cwd: str | Path | None = None,
-        shell: bool | None = None,
-        env: Mapping[str, str] | None = None,
-        nice: int | CpuPriority | None = None,
-        restore_terminal: bool | None = None,
-        auto_run: bool = True,
-    ) -> None:
-        resolved_mode = InteractiveMode(mode)
-        if resolved_mode is InteractiveMode.PSEUDO_TERMINAL:
-            raise ValueError("Use PseudoTerminalProcess for pseudo_terminal mode")
-
-        command, shell = _normalize_command(command, shell)
-        self.command = command
-        self.shell = shell
-        self.cwd = str(cwd) if cwd is not None else None
-        self.env = dict(env) if env is not None else os.environ.copy()
-        self.nice = normalize_nice(nice)
-        self.launch_spec = interactive_launch_spec(resolved_mode)
-        self.restore_terminal = (
-            self.launch_spec.restore_terminal
-            if restore_terminal is None
-            else restore_terminal
-        )
-        self._proc: NativeProcess | None = None
-        self._end_time: float | None = None
-        self._finalized = False
-        self.exit_reason: str | None = None
-        self.interrupt_count = 0
-        self.interrupted_by_caller = False
-        self._exit_status: ExitStatus | None = None
-
-        if auto_run:
-            self.start()
-
-    def start(self) -> None:
-        if self._proc is not None:
-            raise RuntimeError("Interactive process already started")
-        creationflags = self.launch_spec.creationflags if sys.platform == "win32" else None
-        self._proc = NativeProcess(
-            self.command,
-            cwd=self.cwd,
-            env=self.env,
-            shell=self.shell,
-            capture=False,
-            creationflags=creationflags,
-            nice=self.nice,
-            create_process_group=(
-                sys.platform != "win32"
-                and self.launch_spec.mode is InteractiveMode.CONSOLE_ISOLATED
-            ),
-        )
-        self._proc.start()
-
-    def poll(self) -> int | None:
-        if self._proc is None:
-            return None
-        return self._proc.poll()
-
-    def wait(self, timeout: float | None = None, *, raise_on_abnormal_exit: bool = False) -> int:
-        if self._proc is None:
-            raise RuntimeError("Interactive process is not running")
-        try:
-            code = self._proc.wait(timeout=timeout)
-        except TimeoutError as exc:
-            self.kill()
-            self._finalize("timeout")
-            raise TimeoutError("Interactive process timed out") from exc
-        self._finalize("exit")
-        self._exit_status = classify_exit_status(code, KEYBOARD_INTERRUPT_EXIT_CODES)
-        if code in KEYBOARD_INTERRUPT_EXIT_CODES:
-            raise KeyboardInterrupt
-        if raise_on_abnormal_exit and self._exit_status.abnormal:
-            raise ProcessAbnormalExit(self._exit_status)
-        return code
-
-    def terminate(self) -> None:
-        if self._proc is None:
-            raise RuntimeError("Interactive process is not running")
-        if self.poll() is not None:
-            self._finalize("exit")
-            return
-        if self.launch_spec.mode is InteractiveMode.CONSOLE_ISOLATED:
-            self._proc.terminate_group()
-        else:
-            self._proc.terminate()
-        self._wait_for_exit()
-        self._finalize("terminate")
-
-    def kill(self) -> None:
-        if self._proc is None:
-            raise RuntimeError("Interactive process is not running")
-        if self.poll() is not None:
-            self._finalize("exit")
-            return
-        if self.launch_spec.mode is InteractiveMode.CONSOLE_ISOLATED:
-            self._proc.kill_group()
-        else:
-            self._proc.kill()
-        self._wait_for_exit()
-        self._finalize("kill")
-
-    def close(self) -> None:
-        if self._proc is None or self._finalized:
-            return
-        with suppress(*_PTY_CLEANUP_ERRORS):
-            if self.poll() is None:
-                self.kill()
-                return
-            self._finalize("exit")
-
-    def __del__(self) -> None:
-        with suppress(*_PTY_CLEANUP_ERRORS):
-            self.close()
-
-    def send_interrupt(self) -> None:
-        if self._proc is None:
-            raise RuntimeError("Interactive process is not running")
-        self.interrupt_count += 1
-        self.interrupted_by_caller = True
-        if (
-            sys.platform != "win32"
-            and self.launch_spec.mode is InteractiveMode.CONSOLE_ISOLATED
-            and self.pid is not None
-        ):
-            with suppress(OSError, AttributeError):
-                os.killpg(self.pid, signal.SIGINT)
-                return
-        self._proc.send_interrupt()
-
-    @property
-    def pid(self) -> int | None:
-        if self._proc is None:
-            return None
-        return self._proc.pid
-
-    @property
-    def exit_status(self) -> ExitStatus | None:
-        code = self.poll()
-        if code is None:
-            return None
-        if self._exit_status is None:
-            self._exit_status = classify_exit_status(code, KEYBOARD_INTERRUPT_EXIT_CODES)
-        return self._exit_status
-
-    def _finalize(self, reason: str) -> None:
-        if self._finalized:
-            return
-        self._finalized = True
-        self._end_time = time.time()
-        self.exit_reason = (
-            "interrupt" if reason == "exit" and self.interrupted_by_caller else reason
-        )
-
-    def _wait_for_exit(self) -> None:
-        try:
-            self._proc.wait(timeout=2)
-        except TimeoutError:
-            self._proc.kill()
-            self._proc.wait(timeout=2)
-
-
-def _windows_pty_command(command: str | list[str], shell: bool) -> list[str]:
-    if shell:
-        if isinstance(command, str):
-            return ["cmd", "/C", command]
-        return ["cmd", "/C", render_command_list(command)]
-    if isinstance(command, str):
-        return [command]
-    return command
-
-
-def _posix_pty_command(
-    command: str | list[str], shell: bool, nice: int | None = None
-) -> list[str]:
-    if shell:
-        if isinstance(command, str):
-            argv = ["sh", "-lc", command]
-        else:
-            argv = ["sh", "-lc", shlex.join(command)]
-    elif isinstance(command, str):
-        argv = [command]
-    else:
-        argv = command
-    if nice is None:
-        return argv
-    return _wrap_posix_pty_command_with_nice(argv, nice)
-
-
-def _wrap_posix_pty_command_with_nice(argv: list[str], nice: int) -> list[str]:
-    return [
-        sys.executable,
-        "-c",
-        (
-            "import os, sys\n"
-            "os.setpriority(os.PRIO_PROCESS, 0, int(sys.argv[1]))\n"
-            "os.execvp(sys.argv[2], sys.argv[2:])\n"
-        ),
-        str(nice),
-        *argv,
-    ]
-
-
-def _pty_command(command: str | list[str], shell: bool, nice: int | None = None) -> list[str]:
-    if sys.platform == "win32":
-        return _windows_pty_command(command, shell)
-    return _posix_pty_command(command, shell, nice)
-
-
-def _normalize_command(
-    command: str | list[str], shell: bool | None
-) -> tuple[str | list[str], bool]:
-    if isinstance(command, list):
-        return command, bool(shell)
-
-    if shell is True:
-        return command, True
-
-    if shell is False:
-        return _split_command(command), False
-
-    if _contains_shell_metacharacters(command):
-        return command, True
-    return _split_command(command), False
-
-
-def _contains_shell_metacharacters(command: str) -> bool:
-    shell_meta = {"&&", "||", "|", ";", ">", "<", "&"}
-    return any(token in command for token in shell_meta)
-
-
-def _split_command(command: str) -> list[str]:
-    parts = shlex.split(command, posix=False)
-    return [_strip_wrapping_quotes(part) for part in parts]
-
-
-def _apply_process_nice(pid: int | None, nice: int | None) -> None:
-    if pid is None or nice is None:
-        return
-    with suppress(RuntimeError):
-        native_apply_process_nice(pid, nice)
-
-
-def _strip_wrapping_quotes(value: str) -> str:
-    if len(value) >= 2 and value[0] == value[-1] and value[0] in {"'", '"'}:
-        return value[1:-1]
-    return value
-
-
-def interactive_launch_spec(mode: InteractiveMode | str) -> InteractiveLaunchSpec:
-    resolved = InteractiveMode(mode)
-    if resolved is InteractiveMode.PSEUDO_TERMINAL:
-        return InteractiveLaunchSpec(
-            mode=resolved,
-            uses_pty=True,
-            ctrl_c_owner="child",
-            creationflags=None,
-            restore_terminal=True,
-        )
-    if resolved is InteractiveMode.CONSOLE_ISOLATED:
-        return InteractiveLaunchSpec(
-            mode=resolved,
-            uses_pty=False,
-            ctrl_c_owner="parent",
-            creationflags=CREATE_NEW_PROCESS_GROUP if sys.platform == "win32" else None,
-            restore_terminal=True,
-        )
-    return InteractiveLaunchSpec(
-        mode=resolved,
-        uses_pty=False,
-        ctrl_c_owner="shared",
-        creationflags=None,
-        restore_terminal=False,
-    )
