+++ src/running_process/running_process/_core.py
@@ -1,23 +1,17 @@
 from __future__ import annotations
 
 import os
-import re
-import shlex
 import sys
 import time
 from collections.abc import Callable
 from contextlib import suppress
-from dataclasses import dataclass
-from datetime import datetime, timezone
-from io import TextIOBase
 from pathlib import Path
-from typing import Any, ClassVar, NamedTuple
+from typing import Any, ClassVar
 
 from running_process._native import NativeProcess
 from running_process.command_render import list2cmdline
 from running_process.compat import (
     CREATE_NEW_PROCESS_GROUP,
-    DEVNULL,
     PIPE,
     STDOUT,
     CalledProcessError,
@@ -51,257 +45,29 @@ from running_process.pty import (
     WaitForResult,
     interactive_launch_spec,
 )
+from running_process.running_process._helpers import (
+    _expect_pattern_spec,
+    _make_timestamped_callback,
+    _parse_shebang_command,
+    _safe_console_write,
+    _stdin_mode,
+    _validate_echo_flag,
+    _validate_echo_timestamps,
+    _validate_expect_stream,
+)
+from running_process.running_process._iter import _RunningProcessOutputIterator
+from running_process.running_process._types import (
+    _FINALIZER_CLEANUP_ERRORS,
+    EOS,
+    CapturedProcessStream,
+    EchoCallback,
+    EchoValue,
+    EndOfStream,
+    ProcessInfo,
+)
 from running_process.running_process_manager import RunningProcessManagerSingleton
 
 _BUFSIZE_NOT_SET = object()
-
-
-class EndOfStream:
-    def __repr__(self) -> str:
-        return "EOS"
-
-
-EOS = EndOfStream()
-_FINALIZER_CLEANUP_ERRORS = (OSError, RuntimeError, TimeoutError, ValueError, AttributeError)
-
-
-@dataclass
-class ProcessInfo:
-    pid: int
-    command: str | list[str]
-    duration: float
-
-
-EchoValue = str | bytes
-EchoCallback = Callable[[str], None]
-
-
-def _is_eos(value: object) -> bool:
-    return isinstance(value, EndOfStream)
-
-
-class ProcessOutputEvent(NamedTuple):
-    stdout: EchoValue | EndOfStream | None
-    stderr: EchoValue | EndOfStream | None
-    exit_code: int | None
-
-    @property
-    def streams_drained(self) -> bool:
-        return (self.stdout is None or _is_eos(self.stdout)) and (
-            self.stderr is None or _is_eos(self.stderr)
-        )
-
-    @property
-    def finished_and_drained(self) -> bool:
-        return self.streams_drained and self.exit_code is not None
-
-
-class _RunningProcessOutputIterator:
-    def __init__(self, process: RunningProcess, timeout: float | None) -> None:
-        self._process = process
-        self._timeout = timeout
-        self._streams_drained = False
-        self._finished = False
-
-    def __iter__(self) -> _RunningProcessOutputIterator:
-        return self
-
-    def __next__(self) -> ProcessOutputEvent:
-        if self._finished:
-            raise StopIteration
-        if self._process._pty_process is not None:
-            raise NotImplementedError(
-                "stdout/stderr tuple iteration is only available for pipe-backed RunningProcess"
-            )
-        if not self._process.capture:
-            raise NotImplementedError("stdout/stderr tuple iteration requires capture=True")
-        if self._streams_drained:
-            exit_code = self._process.poll()
-            if exit_code is None:
-                exit_code = self._process._proc.wait(timeout=self._timeout)
-                self._process._end_time = self._process._end_time or time.time()
-                RunningProcessManagerSingleton.unregister(self._process)
-            self._finished = True
-            return ProcessOutputEvent(EOS, EOS, exit_code)
-
-        status, stream, line = self._process._proc.take_combined_line(self._timeout)
-        if status == "timeout":
-            raise TimeoutError("No stdout or stderr available before timeout")
-        if status == "line" and stream is not None and line is not None:
-            exit_code = self._process.returncode
-            if stream == "stdout":
-                return ProcessOutputEvent(self._process._format(line), None, exit_code)
-            return ProcessOutputEvent(None, self._process._format(line), exit_code)
-        exit_code = self._process.poll()
-
-        self._streams_drained = True
-        if exit_code is None:
-            try:
-                grace_timeout = 0.01
-                if self._timeout is not None:
-                    grace_timeout = min(self._timeout, grace_timeout)
-                exit_code = self._process._proc.wait(timeout=grace_timeout)
-                self._process._end_time = self._process._end_time or time.time()
-                RunningProcessManagerSingleton.unregister(self._process)
-            except TimeoutError:
-                exit_code = None
-        if exit_code is not None:
-            self._finished = True
-        return ProcessOutputEvent(EOS, EOS, exit_code)
-
-
-class CapturedProcessStream:
-    def __init__(self, process: RunningProcess, stream: str) -> None:
-        self._process = process
-        self._stream = stream
-
-    def available(self) -> bool:
-        if self._stream == "combined":
-            return self._process.has_pending_output()
-        if self._stream == "stdout":
-            return self._process.has_pending_stdout()
-        return self._process.has_pending_stderr()
-
-    def read(self) -> str | bytes:
-        return self._process._captured_stream_value(self._stream)
-
-    def drain(self) -> list[EchoValue] | list[tuple[str, EchoValue]]:
-        if self._stream == "stdout":
-            return self._process.drain_stdout()
-        if self._stream == "stderr":
-            return self._process.drain_stderr()
-        return self._process.drain_combined()
-
-    def __repr__(self) -> str:
-        return repr(self.read())
-
-    def __str__(self) -> str:
-        value = self.read()
-        if isinstance(value, str):
-            return value
-        return value.decode(self._process.encoding, self._process.errors)
-
-    def __bytes__(self) -> bytes:
-        value = self.read()
-        if isinstance(value, bytes):
-            return value
-        return value.encode(self._process.encoding, self._process.errors)
-
-    def __eq__(self, other: object) -> bool:
-        return self.read() == other
-
-    def __bool__(self) -> bool:
-        return bool(self.read())
-
-    def __len__(self) -> int:
-        return len(self.read())
-
-    def __contains__(self, item: object) -> bool:
-        value = self.read()
-        return item in value  # type: ignore[operator]
-
-    def __getattr__(self, name: str) -> Any:
-        return getattr(self.read(), name)
-
-
-def _safe_console_write(stream: TextIOBase, line: EchoValue) -> None:
-    text = line.decode("utf-8", errors="replace") if isinstance(line, bytes) else line
-    try:
-        stream.write(text)
-        stream.write("\n")
-    except UnicodeEncodeError:
-        encoding = stream.encoding or "utf-8"
-        rendered = text.encode(encoding, errors="replace")
-        if hasattr(stream, "buffer"):
-            stream.buffer.write(rendered + b"\n")
-        else:
-            stream.write(rendered.decode(encoding, errors="replace"))
-            stream.write("\n")
-    stream.flush()
-
-
-def _stdin_mode(stdin: int | Any | None, has_input: bool) -> str:
-    if has_input:
-        return "piped"
-    if stdin is None:
-        return "inherit"
-    if stdin is DEVNULL:
-        return "null"
-    if stdin is PIPE:
-        return "piped"
-    raise ValueError("unsupported stdin value for RunningProcess; use None, PIPE, or DEVNULL")
-
-
-def _validate_echo_flag(echo: bool | EchoCallback) -> None:
-    if not isinstance(echo, bool) and not callable(echo):
-        raise TypeError(f"echo must be bool or callable, got {type(echo).__name__}")
-
-
-def _validate_echo_timestamps(echo_timestamps: str | None) -> None:
-    if echo_timestamps is not None and echo_timestamps not in ("relative", "absolute"):
-        raise ValueError(
-            f"echo_timestamps must be None, 'relative', or 'absolute', got {echo_timestamps!r}"
-        )
-
-
-def _make_timestamped_callback(
-    inner: EchoCallback,
-    mode: str,
-    start_time: float,
-) -> EchoCallback:
-    if mode == "relative":
-
-        def _relative_cb(line: str) -> None:
-            elapsed = time.time() - start_time
-            inner(f"[{elapsed:.2f}] {line}")
-
-        return _relative_cb
-
-    def _absolute_cb(line: str) -> None:
-        stamp = datetime.now(timezone.utc).strftime("%H:%M:%S.%f")[:-3]
-        inner(f"[{stamp}] {line}")
-
-    return _absolute_cb
-
-
-def _parse_shebang_command(script_path: Path) -> list[str]:
-    first_line = script_path.read_text(encoding="utf-8", errors="replace").splitlines()[0]
-    if first_line.startswith("\ufeff"):
-        first_line = first_line.removeprefix("\ufeff")
-    if not first_line.startswith("#!"):
-        raise ValueError(f"Script does not start with a shebang: {script_path}")
-
-    parts = shlex.split(first_line[2:].strip(), posix=True)
-    if not parts:
-        raise ValueError(f"Invalid shebang in script: {script_path}")
-
-    interpreter = parts[0]
-    if ("/" in interpreter or "\\" in interpreter) and not Path(interpreter).exists():
-        parts[0] = Path(interpreter).name
-
-    if Path(parts[0]).name == "env":
-        env_args = parts[1:]
-        if env_args and env_args[0] in {"-S", "--split-string"}:
-            env_args = env_args[1:]
-        if not env_args:
-            raise ValueError(f"Shebang env launcher has no command: {script_path}")
-        parts = env_args
-
-    return parts
-
-
-def _validate_expect_stream(stream: str) -> str:
-    if stream not in {"stdout", "stderr", "combined"}:
-        raise ValueError("stream must be 'stdout', 'stderr', or 'combined'")
-    return stream
-
-
-def _expect_pattern_spec(pattern: ExpectPattern) -> tuple[str, bool]:
-    if isinstance(pattern, str):
-        return pattern, False
-    if isinstance(pattern, re.Pattern):
-        return pattern.pattern, True
-    raise TypeError("pattern must be a string or compiled regex pattern")
 
 
 class RunningProcess:
@@ -1387,27 +1153,3 @@ class RunningProcess:
             if deadline is not None and time.time() >= deadline:
                 process._handle_timeout(timeout)
             time.sleep(0.01)
-
-
-def subprocess_run(
-    command: str | list[str],
-    cwd: Path | None,
-    check: bool,
-    timeout: int,
-    on_timeout: Callable[[ProcessInfo], None] | None = None,
-    nice: int | CpuPriority | None = None,
-) -> CompletedProcess[str]:
-    try:
-        return RunningProcess.run(
-            command,
-            cwd=cwd,
-            check=check,
-            timeout=timeout,
-            capture_output=True,
-            on_timeout=on_timeout,
-            nice=nice,
-        )
-    except TimeoutExpired as exc:
-        raise RuntimeError(
-            f"CRITICAL: Process timed out after {timeout} seconds: {command}"
-        ) from exc
